Skipped frames logic moved to JAABA plugin, testing.

+++ src/gui/image_dispatcher.hpp
@@ -95,7 +95,7 @@ namespace bias
             std::shared_ptr<TestConfig>testConfig_;
             std::shared_ptr<Lockable<GetTime>> gettime_;
             std::vector<unsigned int>queue_size;
-            std::vector<int64_t>time_stamps1;
+            std::vector<float>time_stamps1;
             std::vector<std::vector<int64_t>> time_stamps;
     };
